Fix Mysql - Sqlite (#16)
* Update nombres automaticos de indices de las migraciones (compatibilidad sqlite).

* Change cambio nombre app

---------

+++ database/migrations/MER/2025_12_09_231636_create_documentos_vehiculo_table.php
@@ -4,15 +4,14 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::create('documentos_vehiculo', function (Blueprint $table) {
-            $table->integer('id')->primary();
+            $table->unsignedBigInteger('id')->primary();
             $table->smallinteger('idtipdocveh')->index('idtipdocveh');
             $table->string('numdoc', 45);
             $table->string('empexp', 150);
